fix analyze issues
