+++ examples/lib/stories/games/trex/background/cloud.dart
@@ -2,8 +2,9 @@ import 'package:flame/components.dart';
 
 import '../random_extension.dart';
 import '../trex_game.dart';
+import 'cloud_manager.dart';
 
-class Cloud extends SpriteComponent {
+class Cloud extends SpriteComponent with ParentIsA<CloudManager> {
   Cloud({required Vector2 position})
       : cloudGap = random.fromRange(
           minCloudGap,
@@ -35,7 +36,7 @@ class Cloud extends SpriteComponent {
     if (shouldRemove) {
       return;
     }
-    x -= (parent as CloudManager).cloudSpeed.ceil() * 50 * dt;
+    x -= parent.cloudSpeed.ceil() * 50 * dt;
 
     if (!isVisible) {
       removeFromParent();
@@ -52,42 +53,5 @@ class Cloud extends SpriteComponent {
     y = ((absolutePosition.y / 2 - (maxSkyLevel - minSkyLevel)) +
             random.fromRange(minSkyLevel, maxSkyLevel)) -
         absolutePositionOf(absoluteTopLeftPosition).y;
-  }
-}
-
-class CloudManager extends PositionComponent with HasGameRef<TRexGame> {
-  final double cloudFrequency = 0.5;
-  final int maxClouds = 20;
-  final double bgCloudSpeed = 0.2;
-
-  void addCloud() {
-    final cloudPosition = Vector2(
-      gameRef.size.x + Cloud.initialSize.x + 10,
-      ((absolutePosition.y / 2 - (Cloud.maxSkyLevel - Cloud.minSkyLevel)) +
-              random.fromRange(Cloud.minSkyLevel, Cloud.maxSkyLevel)) -
-          absolutePosition.y,
-    );
-    add(Cloud(position: cloudPosition));
-  }
-
-  double get cloudSpeed => bgCloudSpeed / 1000 * gameRef.currentSpeed;
-
-  @override
-  void update(double dt) {
-    super.update(dt);
-    final int numClouds = children.length;
-    if (numClouds > 0) {
-      final lastCloud = children.last as Cloud;
-      if (numClouds < maxClouds &&
-          (gameRef.size.x / 2 - lastCloud.x) > lastCloud.cloudGap) {
-        addCloud();
-      }
-    } else {
-      addCloud();
-    }
-  }
-
-  void reset() {
-    removeAll(children);
   }
 }
